[pinephone-1.1] Try to load hinted dtb (fdtfile), fallback to /sun50i-a64-pinephone-1.1.dtb
Newer u-boots will properly set ${fdtfile} on production PinePhones to hint
the correct board (currently the choice is between 1.1 "Braveheart" and 1.2 "Community
Edition").

This commit modifies the boot script to try to load the hinted devicetree, and fallbacking
to /sun50i-a64-pinephone-1.1.dtb as it did before.

The devicetrees must be into the "/allwinner" directory, as u-boot will use that prefix while
setting the ${fdtfile} variable in the environment. See [0] for more details.

[0] https://gitlab.com/pine64-org/u-boot/-/commit/16d23194ec7eb3663bca1032eff1dc2b134cd4f8

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>